Admin-Dashboard/backend/requirements.txt
fastapi==0.116.1
sqlalchemy==2.0.43
psycopg2-binary==2.9.10
python-dotenv==1.1.1
